Bugfix: main function deletes deleted project dir
the main function try to delete "project_path" directory which is already deleted by find_strings() function.

+++ truffleHog/truffleHog.py
@@ -56,7 +56,8 @@ def main():
     do_entropy = str2bool(args.do_entropy)
     output = find_strings(args.git_url, args.since_commit, args.max_depth, args.output_json, args.do_regex, do_entropy, surpress_output=False, branch=args.branch, repo_path=args.repo_path)
     project_path = output["project_path"]
-    shutil.rmtree(project_path, onerror=del_rw)
+    if os.path.isdir(project_path):
+        shutil.rmtree(project_path, onerror=del_rw)
     if args.cleanup:
         clean_up(output)
     if output["foundIssues"]:
